+++ fvdb/viz/__init__.py
@@ -3,7 +3,7 @@
 #
 from ._camera_view import CamerasView
 from ._fog_volume_view import FogVolumeView
-from ._gaussian_splat_3d_view import GaussianSplat3dView, ShOrderingMode
+from ._gaussian_splat_3d_view import GaussianSplat3dView
 from ._image_view import ImageView
 from ._level_set_view import LevelSetView
 from ._point_cloud_view import PointCloudView
@@ -21,7 +21,6 @@ __all__ = [
     "ImageView",
     "LevelSetView",
     "get_scene",
-    "ShOrderingMode",
     "PointCloudView",
     "Scene",
     "grid_edge_network",